Updated devDoc.txt with some TODO's that I need to do to help with the new processLogs.py file.
Updated index.js so that it is logging less.  A lot of the logging is for
debugging, so I added a flag DEBUG_MODE.  Set to true to enable the
debugging logging.

processLogs.py can take downloaded log files from aws and unzip them,
concatenate them as well as search through the logs to figure out how many
times each function is called.  More features to come.

Run this on the aws cli to get the log files after exporting them from the
cloudwatch.
Make sure you're in the toos/logs folder.
	aws s3 cp s3://resistorassistantlogs . --recursive.