Rearranged git so that different versions are in different foilders, instead of creating a new branch for each version.
I can see some issues with this way of doing things.  Since the repository
is all alexa apps, if I want to create a branch for implementing a
function or fixing a bug.   Instead of using the branching feature, I need
to create a dev branch folder and make my changes in there.  Then merge
back to the trunk.

It seems like the other option is to create a repository for each project
and use the branching within each repository.  It seems like if I want to
go back to a specific version of the app, this would make it harder.  I
would also have a bunch more folders on the server to keep track of all
the bare repositories.

I need to look into this more, but I need to do something so this doesn't
stomp on the new alexa app I'm working on.